+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/EdcConfiguration.java
@@ -92,6 +92,8 @@ public class EdcConfiguration {
         private Duration requestTtl;
 
         private String urnPrefix;
+
+        private String submodelSuffix;
     }